Message/Optimization changes by Karsten from 2000/08/05 (plus .cvsignore updates, most notably at Objects directory  all generated files are added to it)

+++ Atlas/Message/MEncoder.cpp
@@ -18,7 +18,7 @@ Encoder::Encoder(Atlas::Bridge* b)
 void Encoder::listItem(const Object& obj)
 {
     switch (obj.getType()) {
-        case Object::TYPE_INT: b->listItem(obj.asInt()); break;
+        case Object::TYPE_INT: b->listItem((int)obj.asInt()); break;
         case Object::TYPE_FLOAT: b->listItem(obj.asFloat()); break;
         case Object::TYPE_STRING: b->listItem(obj.asString()); break;
         case Object::TYPE_MAP: {
@@ -46,7 +46,7 @@ void Encoder::listItem(const Object& obj)
 void Encoder::mapItem(const string& name, const Object& obj)
 {
     switch (obj.getType()) {
-        case Object::TYPE_INT: b->mapItem(name, obj.asInt()); break;
+        case Object::TYPE_INT: b->mapItem(name, (int)obj.asInt()); break;
         case Object::TYPE_FLOAT: b->mapItem(name, obj.asFloat()); break;
         case Object::TYPE_STRING: b->mapItem(name, obj.asString()); break;
         case Object::TYPE_MAP: {
